+++ packages/ui/src/icon-menu/_components/index.ts
@@ -1,6 +1,5 @@
 export * from "./color-picker";
 export * from "./menu-search-bar";
 export * from "./upload-form";
-export * from "./use-default-factories";
 export * from "./virtualized-icon-grid";
 export type * from "./types";
