--- packages/ui/src/icon-menu/_components/use-default-factories.ts
@@ -1,13 +1,0 @@
-import {
-  useEmojiFactory,
-  useLucideFactory,
-  useUploadFactory,
-} from "../factories";
-import type { IconFactoryResult } from "../factories";
-
-export function useDefaultFactories(): IconFactoryResult[] {
-  const emoji = useEmojiFactory();
-  const lucide = useLucideFactory();
-  const upload = useUploadFactory();
-  return [emoji, lucide, upload];
-}
